TanayLabUtilities

TanayLabUtilities.TanayLabUtilities — Module

Generic utilities used by the Julia code for the Tanay lab.

This is a somewhat arbitrary collection of "generally useful" functions, created on an as-needed bases. No claim is made for true universality or for suitability for a particular purpose.

If you write using TanayLabUtilities everything will be exported into your global namespaces. Since this is a collection of loosely related functions, you may prefer to write using TanayLabUtilities.SomethingSpecific instead, or just import the specific symbol(s) you need.

The included modules are:

Index

TanayLabUtilities.BitsLoops

TanayLabUtilities.Brief

TanayLabUtilities.CachedIsPath

TanayLabUtilities.ChiSquared

TanayLabUtilities.CrossValidations

TanayLabUtilities.Documentation

TanayLabUtilities.Downsample

TanayLabUtilities.ExtendedReadWriteLocks

TanayLabUtilities.FlameTime

TanayLabUtilities.GlobalWeakCache

TanayLabUtilities.GroupedCorrelations

TanayLabUtilities.Handlers

TanayLabUtilities.KMeans

TanayLabUtilities.Logger

TanayLabUtilities.MatrixFormats

TanayLabUtilities.MatrixLayouts

TanayLabUtilities.MostlyReadWriteLocks

TanayLabUtilities.OldLinuxMmap

TanayLabUtilities.ParallelDistances

TanayLabUtilities.ParallelLoops

TanayLabUtilities.ParallelStorage

TanayLabUtilities.Permute

TanayLabUtilities.ReadOnlyArrays

TanayLabUtilities.SparseStatistics

TanayLabUtilities.TanayLabUtilities

TanayLabUtilities.Types

TanayLabUtilities.UniqueNames

TanayLabUtilities.ZeroCorrelation

TanayLabUtilities.Brief.MAX_BRIEF_STRING

TanayLabUtilities.CachedIsPath.IS_PATH_CACHE_TIMEOUT

TanayLabUtilities.Documentation.DEFAULT

TanayLabUtilities.MatrixLayouts.Columns

TanayLabUtilities.MatrixLayouts.GLOBAL_INEFFICIENT_ACTION_HANDLER

TanayLabUtilities.MatrixLayouts.Rows

TanayLabUtilities.CrossValidations.CrossValidationIndices

TanayLabUtilities.ExtendedReadWriteLocks.ExtendedReadWriteLock

TanayLabUtilities.GroupedCorrelations.GroupedSeriesCorrelations

TanayLabUtilities.Handlers.AbnormalHandler

TanayLabUtilities.KMeans.KMeansBuffers

TanayLabUtilities.KMeans.KmeansResultView

TanayLabUtilities.MostlyReadWriteLocks.MostlyReadWriteLock

TanayLabUtilities.ParallelStorage.ReusableStorage

TanayLabUtilities.ReadOnlyArrays.ReadOnlyArray

TanayLabUtilities.Types.Maybe

TanayLabUtilities.Types.Unsure

TanayLabUtilities.Brief.brief

TanayLabUtilities.Brief.delimited_number

TanayLabUtilities.Brief.percent

TanayLabUtilities.CachedIsPath.cached_ispath

TanayLabUtilities.CachedIsPath.report_modified!

TanayLabUtilities.ChiSquared.chi_squared

TanayLabUtilities.CrossValidations.pick_cross_validation_indices

TanayLabUtilities.Documentation.function_default

TanayLabUtilities.Downsample.downsample

TanayLabUtilities.Downsample.downsamples

TanayLabUtilities.Downsample.round_counts

TanayLabUtilities.ExtendedReadWriteLocks.has_read_lock

TanayLabUtilities.ExtendedReadWriteLocks.has_write_lock

TanayLabUtilities.ExtendedReadWriteLocks.lock_write

TanayLabUtilities.ExtendedReadWriteLocks.trylock_write

TanayLabUtilities.ExtendedReadWriteLocks.unlock_write

TanayLabUtilities.FlameTime.finalize_flameview

TanayLabUtilities.FlameTime.flame_timed

TanayLabUtilities.GlobalWeakCache.get_through_global_weak_cache

TanayLabUtilities.GroupedCorrelations.correlation_per_series!

TanayLabUtilities.GroupedCorrelations.mean_correlation

TanayLabUtilities.GroupedCorrelations.mean_correlation_if_group_replaced

TanayLabUtilities.GroupedCorrelations.replace_group!

TanayLabUtilities.Handlers.handle_abnormal

TanayLabUtilities.KMeans.kmeans_in_buffers

TanayLabUtilities.KMeans.kmeans_in_buffers!

TanayLabUtilities.KMeans.kmeans_in_rounds

TanayLabUtilities.Logger.setup_logger

TanayLabUtilities.MatrixFormats.base_array

TanayLabUtilities.MatrixFormats.bestify

TanayLabUtilities.MatrixFormats.colptr

TanayLabUtilities.MatrixFormats.copy_array

TanayLabUtilities.MatrixFormats.dense_mask_vector

TanayLabUtilities.MatrixFormats.densify

TanayLabUtilities.MatrixFormats.embed_dense_matrix_in_sparse_matrix

TanayLabUtilities.MatrixFormats.embed_sparse_matrix_in_sparse_matrix

TanayLabUtilities.MatrixFormats.indtype_for_size

TanayLabUtilities.MatrixFormats.nzind

TanayLabUtilities.MatrixFormats.nzval

TanayLabUtilities.MatrixFormats.rowval

TanayLabUtilities.MatrixFormats.similar_array

TanayLabUtilities.MatrixFormats.sparse_mask_vector

TanayLabUtilities.MatrixFormats.sparse_matrix_csc

TanayLabUtilities.MatrixFormats.sparse_vector

TanayLabUtilities.MatrixFormats.sparsify

TanayLabUtilities.MatrixLayouts.axis_name

TanayLabUtilities.MatrixLayouts.flip

TanayLabUtilities.MatrixLayouts.flipped

TanayLabUtilities.MatrixLayouts.major_axis

TanayLabUtilities.MatrixLayouts.minor_axis

TanayLabUtilities.MatrixLayouts.other_axis

TanayLabUtilities.MatrixLayouts.relayout

TanayLabUtilities.MatrixLayouts.relayout!

TanayLabUtilities.MatrixLayouts.require_major_axis

TanayLabUtilities.MatrixLayouts.require_minor_axis

TanayLabUtilities.OldLinuxMmap.mmap_populate_if_old_linux_ramdisk

TanayLabUtilities.OldLinuxMmap.mmap_with_small_pages

TanayLabUtilities.ParallelDistances.parallel_colwise

TanayLabUtilities.ParallelDistances.parallel_pairwise

TanayLabUtilities.ParallelDistances.parallel_pairwise_closest

TanayLabUtilities.ParallelLoops.DebugProgress

TanayLabUtilities.ParallelLoops.DebugProgressUnknown

TanayLabUtilities.ParallelLoops.is_in_parallel_loop

TanayLabUtilities.ParallelLoops.parallel_loop_with_rng

TanayLabUtilities.ParallelLoops.parallel_loop_wo_rng

TanayLabUtilities.ParallelStorage.get_reusable!

TanayLabUtilities.ParallelStorage.put_reusable!

TanayLabUtilities.ParallelStorage.reset_reusable_storage!

TanayLabUtilities.ParallelStorage.with_reusable

TanayLabUtilities.Permute.permute_dense_matrix_both!

TanayLabUtilities.Permute.permute_dense_matrix_columns!

TanayLabUtilities.Permute.permute_dense_matrix_rows!

TanayLabUtilities.Permute.permute_sparse_matrix_both!

TanayLabUtilities.Permute.permute_sparse_matrix_both_buffers!

TanayLabUtilities.Permute.permute_sparse_matrix_columns!

TanayLabUtilities.Permute.permute_sparse_matrix_columns_buffers!

TanayLabUtilities.Permute.permute_sparse_matrix_rows!

TanayLabUtilities.Permute.permute_sparse_matrix_rows_buffers!

TanayLabUtilities.Permute.permute_sparse_vector!

TanayLabUtilities.Permute.permute_sparse_vector_buffers!

TanayLabUtilities.Permute.permute_vector!

TanayLabUtilities.ReadOnlyArrays.is_read_only_array

TanayLabUtilities.ReadOnlyArrays.mutable_array

TanayLabUtilities.ReadOnlyArrays.read_only_array

TanayLabUtilities.SparseStatistics.auc_per_category

TanayLabUtilities.SparseStatistics.auc_per_group!

TanayLabUtilities.SparseStatistics.sparse_median

TanayLabUtilities.SparseStatistics.sparse_quantile

TanayLabUtilities.SparseStatistics.sparse_std

TanayLabUtilities.SparseStatistics.sparse_var

TanayLabUtilities.UniqueNames.unique_name

TanayLabUtilities.ZeroCorrelation.zero_cor_between_matrices_columns

TanayLabUtilities.ZeroCorrelation.zero_cor_between_matrix_columns

TanayLabUtilities.ZeroCorrelation.zero_cor_between_vector_and_matrix_columns

TanayLabUtilities.ZeroCorrelation.zero_cor_between_vectors

TanayLabUtilities.BitsLoops.@foreach_true_index

TanayLabUtilities.BitsLoops.@foreach_true_index_position

TanayLabUtilities.Documentation.@documented

TanayLabUtilities.Documentation.@source_location

TanayLabUtilities.Logger.@logged

TanayLabUtilities.MatrixLayouts.@assert_matrix

TanayLabUtilities.MatrixLayouts.@assert_vector

TanayLabUtilities.MatrixLayouts.@check_turbo_matrix

TanayLabUtilities.MatrixLayouts.@check_turbo_vector