Locks

TanayLabUtilities.Locks — Module

Generic (reentrant) read-write locks.

These add functionality on top of ConcurrentUtils ; specifically, they allow querying the status of the lock, and querying the status of the lock. We also provide explicit write_* and read_* functions instead of relying on lock to be write_lock . That is, ReentrantReadWriteLock is not an AbstractLock .

read_write_lock = ReentrantReadWriteLock()

@assert !has_write_lock(read_write_lock)
@assert !has_read_lock(read_write_lock)

with_write_lock(read_write_lock, "top_write") do
    @assert has_write_lock(read_write_lock)
    @assert has_read_lock(read_write_lock)

    with_write_lock(read_write_lock, "nested_write") do
        @assert has_write_lock(read_write_lock)
        @assert has_read_lock(read_write_lock)
    end

    with_read_lock(read_write_lock, "nested_read") do
        @assert has_write_lock(read_write_lock)
        @assert has_read_lock(read_write_lock)
    end
end

@assert !has_write_lock(read_write_lock)
@assert !has_read_lock(read_write_lock)

# output

read_write_lock = ReentrantReadWriteLock()

@assert !has_write_lock(read_write_lock)
@assert !has_read_lock(read_write_lock)

with_read_lock(read_write_lock, "top_read") do
    @assert !has_write_lock(read_write_lock)
    @assert has_read_lock(read_write_lock)

    with_read_lock(read_write_lock, "nested_read") do
        @assert !has_write_lock(read_write_lock)
        @assert has_read_lock(read_write_lock)
    end
end

@assert !has_write_lock(read_write_lock)
@assert !has_read_lock(read_write_lock)

# output

read_write_lock = ReentrantReadWriteLock()

@assert !has_write_lock(read_write_lock)
@assert !has_read_lock(read_write_lock)

with_read_lock(read_write_lock, "top_read") do
    @assert !has_write_lock(read_write_lock)
    @assert has_read_lock(read_write_lock)

    with_write_lock(read_write_lock, "nested_write") do
        @assert false
    end
end

# output

ERROR: trying to obtain write lock for: nested_write
while holding read lock:

TanayLabUtilities.Locks.ReentrantReadWriteLock — Type

struct ReentrantReadWriteLock <: AbstractLock ... end

A read-write lock that supports nested calls. You can nest read locks, write locks, and read locks inside write locks. However, you can't nest write locks in read locks.

Write Lock

TanayLabUtilities.Locks.write_lock — Function

write_lock(read_write_lock::ReentrantReadWriteLock, what::Any...)::Nothing

Obtain a write lock. Each call must be matched by write_unlock . It is possible to nest write_lock / write_unlock call pairs.

When a task has a write lock, no other task can have any lock.

The log messages includes what is being locked.

TanayLabUtilities.Locks.write_unlock — Function

write_unlock(read_write_lock::ReentrantReadWriteLock, what::Any)::Nothing

Release a write lock. Each call must matched a call to write_lock . It is possible to nest write_lock / write_unlock call pairs.

The log messages includes what is being unlocked.

TanayLabUtilities.Locks.with_write_lock — Function

with_write_lock(action::Function, read_write_lock::ReentrantReadWriteLock, what::Any...)::Any

Perform an action while holding a write_lock for the read_write_lock , return its result and write_unlock .

TanayLabUtilities.Locks.has_write_lock — Function

has_write_lock(read_write_lock::ReentrantReadWriteLock)::Bool

Return whether the current task has the write lock.

Read Locks

TanayLabUtilities.Locks.read_lock — Function

read_lock(read_write_lock::ReentrantReadWriteLock, what::Any...)::Nothing

Obtain a read lock. Each call must be matched by read_unlock . It is possible to nest read_lock / read_unlock call pairs, even inside write_lock / write_unlock pair(s); however, you can't nest write_lock / write_unlock inside a read_lock / read_unlock pair.

When a task has a read lock, no other task can have a write lock, but other tasks may also have a read lock.

The log messages includes what is being locked.

TanayLabUtilities.Locks.read_unlock — Function

read_unlock(read_write_lock::ReentrantReadWriteLock, what::Any...)::Nothing

Release a read lock. Each call must matched a call to read_lock . It is possible to nest read_lock / read_unlock call pairs.

The log messages includes what is being unlocked.

TanayLabUtilities.Locks.with_read_lock — Function

with_read_lock(action::Function, read_write_lock::ReentrantReadWriteLock, what::Any...)::Any

Perform an action while holding a read_lock for the read_write_lock , return its result and read_unlock .

TanayLabUtilities.Locks.has_read_lock — Function

has_read_lock(read_write_lock::ReentrantReadWriteLock; read_only::Bool = false)::Bool

Return whether the current task has a read lock or the write lock. If read_only is set, then this will only return whether the current task as a read lock.

Index

TanayLabUtilities.Locks

TanayLabUtilities.Locks.ReentrantReadWriteLock

TanayLabUtilities.Locks.has_read_lock

TanayLabUtilities.Locks.has_write_lock

TanayLabUtilities.Locks.read_lock

TanayLabUtilities.Locks.read_unlock

TanayLabUtilities.Locks.with_read_lock

TanayLabUtilities.Locks.with_write_lock

TanayLabUtilities.Locks.write_lock

TanayLabUtilities.Locks.write_unlock