ParallelRNG

TanayLabUtilities.ParallelRNG — Module

Reproducible random number generator for parallel loops. Yes, the TaskLocalRNG is supposed to do this, but, it actually depends on the way tasks are allocated to threads. The implementation here will give the same results regardless of the thread scheduling policy.

TanayLabUtilities.ParallelRNG.parallel_loop_with_rng — Function

parallel_loop_with_rng(
    body::Function,
    size::Maybe{Integer} = nothing;
    policy::Symbol = :greedy,
    seed::Maybe{Integer} = nothing,
    rng::Maybe{AbstractRNG} = nothing
)::Nothing

Run the body in parallel (using the @threads policy ) size times, passing it the iteration index, and a separate rng that is seeded to a reproducible state regardless of the allocation of tasks to threads. A copy of this rng is given to each iteration, after being reset to seed + index for reproducibility. If no seed is specified, it is just sampled rng before the loop starts. If the rng isn't given, then this uses (and sets for each iteration) the default_rng() . In this case passing it to the body is redundant but is still done for consistency.

using Test
using Random

size = 10

function collect_rng(rng::AbstractRNG)::Vector{Float64}
    results = zeros(Float64, size)
    parallel_loop_with_rng(size; rng) do index, rng
        results[index] = rand(rng)
    end
    @test results[1] != results[2]
    return results
end

@test collect_rng(MersenneTwister(1)) == collect_rng(MersenneTwister(1))

function collect_default_rng()::Vector{Float64}
    results = zeros(Float64, size)
    parallel_loop_with_rng(size; seed = 123456, policy = :dynamic) do index, _
        results[index] = rand()
    end
    @test results[1] != results[2]
    return results
end

@test collect_default_rng() == collect_default_rng()

println("OK")

# output

OK

Index

TanayLabUtilities.ParallelRNG

TanayLabUtilities.ParallelRNG.parallel_loop_with_rng