TanayLabUtilities

TanayLabUtilities.TanayLabUtilities — Module

Generic utilities used by the Julia code for the Tanay lab.

This is a somewhat arbitrary collection of "generally useful" functions, created on an as-needed bases. No claim is made for true universality or for suitability for a particular purpose.

If you write using TanayLabUtilities everything will be exported into your global namespaces. Since this is a collection of loosely related functions, you may prefer to write using TanayLabUtilities.SomethingSpecific instead, or just import the specific symbol(s) you need.

The included modules are:

Index

TanayLabUtilities.Brief

TanayLabUtilities.Documentation

TanayLabUtilities.Downsample

TanayLabUtilities.Handlers

TanayLabUtilities.Locks

TanayLabUtilities.Logger

TanayLabUtilities.MatrixFormats

TanayLabUtilities.MatrixLayouts

TanayLabUtilities.ParallelRNG

TanayLabUtilities.ParallelStorage

TanayLabUtilities.ReadOnlyArrays

TanayLabUtilities.TanayLabUtilities

TanayLabUtilities.Types

TanayLabUtilities.UniqueNames

TanayLabUtilities.Brief.MAX_BRIEF_STRING

TanayLabUtilities.Documentation.DEFAULT

TanayLabUtilities.MatrixLayouts.Columns

TanayLabUtilities.MatrixLayouts.GLOBAL_INEFFICIENT_ACTION_HANDLER

TanayLabUtilities.MatrixLayouts.Rows

TanayLabUtilities.Handlers.AbnormalHandler

TanayLabUtilities.Locks.ReentrantReadWriteLock

TanayLabUtilities.ParallelStorage.ReusableStorage

TanayLabUtilities.ReadOnlyArrays.ReadOnlyArray

TanayLabUtilities.Types.Maybe

TanayLabUtilities.Types.Unsure

TanayLabUtilities.Brief.brief

TanayLabUtilities.Brief.percent

TanayLabUtilities.Documentation.function_default

TanayLabUtilities.Downsample.downsample

TanayLabUtilities.Downsample.downsamples

TanayLabUtilities.Handlers.handle_abnormal

TanayLabUtilities.Locks.has_read_lock

TanayLabUtilities.Locks.has_write_lock

TanayLabUtilities.Locks.read_lock

TanayLabUtilities.Locks.read_unlock

TanayLabUtilities.Locks.with_read_lock

TanayLabUtilities.Locks.with_write_lock

TanayLabUtilities.Locks.write_lock

TanayLabUtilities.Locks.write_unlock

TanayLabUtilities.Logger.setup_logger

TanayLabUtilities.MatrixFormats.bestify

TanayLabUtilities.MatrixFormats.colptr

TanayLabUtilities.MatrixFormats.copy_array

TanayLabUtilities.MatrixFormats.dense_mask_vector

TanayLabUtilities.MatrixFormats.densify

TanayLabUtilities.MatrixFormats.indtype_for_size

TanayLabUtilities.MatrixFormats.nzind

TanayLabUtilities.MatrixFormats.nzval

TanayLabUtilities.MatrixFormats.rowval

TanayLabUtilities.MatrixFormats.similar_array

TanayLabUtilities.MatrixFormats.sparse_mask_vector

TanayLabUtilities.MatrixFormats.sparse_matrix_csc

TanayLabUtilities.MatrixFormats.sparse_vector

TanayLabUtilities.MatrixFormats.sparsify

TanayLabUtilities.MatrixLayouts.axis_name

TanayLabUtilities.MatrixLayouts.major_axis

TanayLabUtilities.MatrixLayouts.minor_axis

TanayLabUtilities.MatrixLayouts.other_axis

TanayLabUtilities.MatrixLayouts.relayout

TanayLabUtilities.MatrixLayouts.relayout!

TanayLabUtilities.MatrixLayouts.require_major_axis

TanayLabUtilities.MatrixLayouts.require_minor_axis

TanayLabUtilities.MatrixLayouts.transposer

TanayLabUtilities.ParallelRNG.parallel_loop_with_rng

TanayLabUtilities.ParallelStorage.get_reusable!

TanayLabUtilities.ParallelStorage.put_reusable!

TanayLabUtilities.ParallelStorage.with_reusable

TanayLabUtilities.ReadOnlyArrays.is_read_only_array

TanayLabUtilities.ReadOnlyArrays.mutable_array

TanayLabUtilities.ReadOnlyArrays.read_only_array

TanayLabUtilities.UniqueNames.unique_name

TanayLabUtilities.Documentation.@documented

TanayLabUtilities.Documentation.@source_location

TanayLabUtilities.Logger.@logged

TanayLabUtilities.MatrixLayouts.@assert_matrix

TanayLabUtilities.MatrixLayouts.@assert_vector