Flame Time

TanayLabUtilities.FlameTime — Module

Generate a coarse-grained flamegraph. If the environment variable JULIA_FLAME_MEASUREMENTS_FILE is set, it should contain the path of a file to append measurements into. Every call to flame_timed will append measurement into this file. This file can't be directly viewed using flamegraph.pl . It needs to be converted to a proper flamegraph file using finalize_flamegraph .

Note

Measurements are appended into the file, so that running multiple programs automatically creates a unified file for a whole computational pipeline. If you want to separately measure each program execution, use a different measurements file for each.

This serves as a very lightweight mechanism for visualizing program performance. Every @logged function and every parallel_loop is automatically a measured step, so one gets a pretty decent level of detail "for free". It is possible to supplement this with occasional calls to flame_timed to increase the level of detail. Of course one can ignore all of this and use heavyweight tools such as Profile when profiling for performance in anger.

TanayLabUtilities.FlameTime.flame_timed — Function

flame_timed(body::Function, name::AbstractString)::Any

Add a measurement line to the flamegrah file for the execution of the body under the specified name . The measurement is of the elapsed time for the body in nanoseconds. We distinguish between "serial" measurements (in the top-level main thread) and "parallel" measurements (inside tasks invoked from within a parallel_loop ).

TanayLabUtilities.FlameTime.finalize_flamegraph — Function

finalize_flamegraph(; measurements_path::AbstractString, flamegraph_path::AbstractString)::Nothing

Convert the collected measurements in the measurement_path into a proper flamegraph file, suitable for use with flamegraph.pl and similar tools. This is needed for two reasons: first, the measurements are inclusive (that is, the elapsed time of a step includes the elapsed time of all nested steps). Second, measurements of serial steps (which should add to the total execution time) are mixed with the measurements of parallel steps (which would add up to much more).

Each parallel_loop is therefore measured twice, once as a serial step and once as a parallel step. We scale the parallel time (and everything nested in it) so it adds up to the serial time. We then convert all the inclusive measurements to exclusive ones. The result is a flamegraph that shows the contribution of everything (serial and parallel steps) to the total execution time, under the assumption that the number of iterations in the parallel loops is large enough that it is the sum of parallel steps that matters.

Note

This will not work if you call flame_timed inside an explicit @threads loops. You must use a parallel_loop instead.

Index

TanayLabUtilities.FlameTime

TanayLabUtilities.FlameTime.finalize_flamegraph

TanayLabUtilities.FlameTime.flame_timed